refactor to use DispatchGroup methods
